nambahin configure tailwindcss custom dan menambahkan home page
