+++ src/components/footer.jsx
@@ -1,0 +1,30 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+import logo from '../assets/img/logo.png';
+
+const Footer = () => {
+  return (
+    <footer className="bg-white text-(--black)">
+      {/* Horizontal Line */}
+      <div className="border-t-2 mx-10"></div>
+
+      {/* Footer Content */}
+      <div className="flex justify-between items-center px-10 py-20 ">
+        {/* Left Section */}
+        <div className="flex items-center">
+          <span className="text-2xl font-bold flex items-center">
+            <img src={logo} alt="Logo" className="h-24 w-24 inline-block mr-2" />
+            Adopt House
+          </span>
+        </div>
+
+        {/* Right Section */}
+        <div>
+          <p className="text-sm">@2025 adopthouse.com All rights reserved.</p>
+        </div>
+      </div>
+    </footer>
+  );
+};
+
+export default Footer;
