+++ src/components/home-copomnents/card-item.jsx
@@ -1,0 +1,7 @@
+const CardItem = () => {
+    return (
+        <h1>Hello ini card item</h1>
+    )
+}
+
+export default CardItem;
