+++ src/components/home-copomnents/cards-list.jsx
@@ -1,0 +1,11 @@
+import CardItem from "./card-item";
+
+const CardList = () => {
+    return (
+        <div className="card-list w-full h-screen">
+            <CardItem />
+        </div>
+    )
+}
+
+export default CardList;
