+++ src/index.css
@@ -1,2 +1,59 @@
-@import url("https://fonts.googleapis.com/css2?family=Poppins&display=swap");
+@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;600;700;800&display=swap');
 @import "tailwindcss";
+
+/* src/index.css */
+@theme {
+  --header-height: 3.5rem;
+
+  --blue-sky: #83b4ff;
+  --white: #ffffff;
+  --dark-grey: #d9d9d9;
+  --grey: #eeeeee;
+  --yellow: #fdffe2;
+  --navy: #5a72a0;
+  --black: #1a2130;
+  --container-color: var(--white);
+  --shadow-small-img: drop-shadow(0 4px 16px var(--primary-shadow-color));
+
+  /* Font and typography */
+  --body-font: "Poppins", sans-serif;
+  --biggest-font-size: 2.5rem;
+  --bigger-font-size: 1.5rem;
+  --h1-font-size: 1.5rem;
+  --h2-font-size: 1.25rem;
+  --h3-font-size: 1rem;
+  --normal-font-size: 0.938rem;
+  --small-font-size: 0.813rem;
+  --smaller-font-size: 0.75rem;
+
+  /* Font weight */
+  --font-regular: 400;
+  --font-medium: 500;
+  --font-semi-bold: 600;
+  --font-bold: 700
+
+  /* z index */
+  --z-tooltip: 10;
+  --z-fixed: 100;
+}
+
+@media screen and (min-width: 1150px) {
+  @theme {
+    --biggest-font-size: 4.25rem;
+    --bigger-font-size: 2.5rem;
+    --h1-font-size: 2.25rem;
+    --h2-font-size: 1.5rem;
+    --h3-font-size: 1.25rem;
+    --normal-font-size: 1rem;
+    --small-font-size: 0.875rem;
+    --smaller-font-size: 0.813rem;
+  }
+}
+
+* {
+  font-family: var(--body-font);
+}
+
+body{
+  background-color: var(--grey);
+}
