+++ src/pages/home/home.jsx
@@ -1,0 +1,14 @@
+import CardList from "../../components/home-copomnents/cards-list";
+import HeroSection from "../../components/home-copomnents/hero-section";
+
+const HomePage = () => {
+
+    return (
+        <>
+            <HeroSection />
+            <CardList />
+        </>
+    )
+};
+
+export default HomePage;
