--- tailwind.config.js
@@ -1,13 +1,0 @@
-export default {
-  content: [
-    "./src/**/*.{js,ts,jsx,tsx,html}",
-  ],
-  theme: {
-    extend: {
-      colors: {
-        bluejeans: '#5A72A0',
-      },
-    },
-  },
-  plugins: [],
-}
